+++ tests/test_cases/test_3.py
@@ -1,16 +1,35 @@
 # this page contains all the test cases for the samplePage
 import os
 import random
-
-from dotenv import load_dotenv
-
 from conftest import new_tab
 
+# For validation
+from playwright.sync_api import expect
+
+# Import for beautiful reporting
+from rich.traceback import install
+install()
+
+from dotenv import load_dotenv
 load_dotenv()
 
 # Project URLs
 proj_url = os.getenv("test_url")
 requisition_list_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/list"
+requisition_approve_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/authorizationList"
+requisition_assign_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/assignRequisitions"
+requisition_accept_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/assignedRequisitionShowList"
+tender_initiation_url = proj_url + "/procurementDashboard/myDashboard#!/methodSelection/show"
+direct_purchase_url = proj_url + "/procurementDashboard/myDashboard#!/directPurchase/show"
+direct_purchase_list_url = proj_url + "/procurementDashboard/myDashboard#!/directPurchase/list"
+item_receive_url = proj_url + "/procurementDashboard/myDashboard#!/itemReceive/show"
+vendor_bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/show"
+bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
+
+
+
+
+
 
 
 proj_user = os.getenv("test_user_name")
@@ -43,14 +62,6 @@ from pages.vendor_bill_payable_list import BillList
 from pages.bill_details_information import BillDetails
 
 
-
-# For validation
-from playwright.sync_api import expect
-
-# Import for beautiful reporting
-from rich.traceback import install
-install()
-
 req_num = ''
 approver_id = ''
 approver_id_2 = ''
@@ -60,6 +71,8 @@ bill_approver_id = ''
 purchase_num = ''
 challan_num = str(random.randint(10000,99999))
 bill_num = str(random.randint(10000,99999))
+bill_recommender1 = ''
+bill_recommender2 = ''
 
 
 def test_1_login_to_create_requisition(page):
@@ -132,7 +145,6 @@ def test_6_login_as_approver_and_approve(page):
         timeout=60000  # Increased timeout for login
     )
     r_page = RequisitionApproveList(page)
-    requisition_approve_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/authorizationList"
     r_page.navigate_to_url(requisition_approve_url)
     r_page.get_full_page_screenshot('full_page_screenshot_8')
     r_page.search_requisition(req_num)
@@ -158,7 +170,6 @@ def test_7_find_approver_of_the_requisition_2(page):
     )
 
     r_page = RequisitionList(page)
-    requisition_list_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/list"
     r_page.navigate_to_url(requisition_list_url)
     r_page.get_full_page_screenshot('full_page_screenshot_11')
     r_page.search_requisition(req_num)
@@ -186,7 +197,6 @@ def test_8_login_as_approver_and_approve_2(page):
     )
 
     r_page = RequisitionApproveList(page)
-    requisition_approve_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/authorizationList"
     r_page.navigate_to_url(requisition_approve_url)
     r_page.get_full_page_screenshot('full_page_screenshot_14')
     r_page.search_requisition(req_num)
@@ -213,7 +223,6 @@ def test_9_check_requisition_approved(page):
     )
 
     r_page = RequisitionList(page)
-    requisition_list_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/list"
     r_page.navigate_to_url(requisition_list_url)
     r_page.get_full_page_screenshot('full_page_screenshot_17')
     r_page.search_requisition(req_num)
@@ -241,7 +250,6 @@ def test_10_check_requisition_assign(page):
     )
 
     r_page = AssignRequisition(page)
-    requisition_assign_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/assignRequisitions"
     r_page.navigate_to_url(requisition_assign_url)
     r_page.assigning_person(assigned_person)
     r_page.search_requisition_for_assigning(req_num)
@@ -272,7 +280,6 @@ def test_11_requisition_accept(page):
     )
 
     r_page = RequisitionAcceptList(page)
-    requisition_accept_url = proj_url + "/procurementDashboard/myDashboard#!/requisition/assignedRequisitionShowList"
     r_page.navigate_to_url(requisition_accept_url)
     r_page.search_requisition(req_num)
     r_page.select_all_requisitions()
@@ -285,7 +292,6 @@ def test_12_create_tender_initiation(page):
     print("Test 12: Creating tender initiation...")
     req_num2 = "REQ20250004556"
     t_page = CreateTenderInitiation(page)
-    tender_initiation_url = proj_url + "/procurementDashboard/myDashboard#!/methodSelection/show"
     t_page.navigate_to_url(tender_initiation_url)
     t_page.search_requisition(req_num)
     t_page.select_all_items()
@@ -301,7 +307,6 @@ def test_13_create_direct_purchase(page):
     req_num2 = "REQ20250004556"
     t_page = CreateDirectPurchase(page)
     try:
-        direct_purchase_url = proj_url + "/procurementDashboard/myDashboard#!/directPurchase/show"
         t_page.navigate_to_url(direct_purchase_url)
         t_page.search_vendor(vendor_name)
         t_page.same_delivery_schedule()
@@ -345,7 +350,6 @@ def test_14_approve_direct_purchase(page):
 
     t_page = DirectPurchaseList(page)
     try:
-        direct_purchase_list_url = proj_url + "/procurementDashboard/myDashboard#!/directPurchase/list"
         t_page.navigate_to_url(direct_purchase_list_url)
         t_page.search_purchase_order(purchase_num)
         # t_page.navigate_to_direct_purchase_detail_page(purchase_num)
@@ -375,7 +379,6 @@ def test_15_item_receive(page):
         timeout=60000
     )
     t_page = ItemReceive(page)
-    item_receive_url = proj_url + "/procurementDashboard/myDashboard#!/itemReceive/show"
     t_page.navigate_to_url(item_receive_url)
     t_page.search_order_for_item_receive(purchase_num)
     t_page.set_challan_number(challan_num)
@@ -404,7 +407,6 @@ def test_16_bill_creation_and_submit(page):
         timeout=60000
     )
     t_page = CreateVendorBillPayable(page)
-    vendor_bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/show"
     t_page.navigate_to_url(vendor_bill_payable_url)
     t_page.search_vendor(vendor_name)
     # t_page.select_order_no(purchase_num)
@@ -421,11 +423,12 @@ def test_16_bill_creation_and_submit(page):
     t_page.get_full_page_screenshot('full_page_screenshot_34')
 
     l2_page = BillList(page)
-    bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
+
     l2_page.navigate_to_url(bill_payable_url)
     l2_page.search_bill(bill_num)
     global bill_recommender1
     bill_recommender1 = str(int(l2_page.find_approver_id()))
+    print(f"Bill Recommender 1: {bill_recommender1}")
 
     # logout from the page
     r2_page = MainNavigationBar(page)
@@ -447,36 +450,41 @@ def test_17_vendor_bill_recommender1_approval(page, new_tab):
     )
 
     l2_page = BillList(page)
-    bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
     l2_page.navigate_to_url(bill_payable_url)
     # l2_page.search_bill(bill_num)
     # l2_page.click_on_bill_num(bill_num)
     l2_page.search_bill(bill_num)
-
+    
+    # # Opening new tab
     new_page = new_tab(lambda p:l2_page.click_on_bill_num(bill_num))
-
     b_page = BillDetails(new_page)
-    document_location = r"C:\Users\shamima.sultana\Downloads\upload_file.pdf"
+
+    # # Preparing document location
+    document_location = os.path.join(current_dir, 'utils', 'upload_file.pdf')
     b_page.upload_document(document_location)
+    print(f"Current directory: {current_dir}")
+    print(f"Document directory: {document_location}")
+
+    # #  Continuing rest of the test
     b_page.get_full_page_screenshot('full_page_screenshot_36')
     b_page.select_bill_type("Regular")
     b_page.get_full_page_screenshot('full_page_screenshot_37')
     b_page.approve_bill()
     b_page.get_full_page_screenshot('full_page_screenshot_38')
 
+    # # Closing new tab
     new_page.close()
 
+    # # Continuing rest of the test in parent tab
     l3_page = BillList(page)
-    bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
     l3_page.navigate_to_url(bill_payable_url)
     # l2_page.search_bill(bill_num)
     l3_page.search_bill(bill_num)
     # b_page.get_full_page_screenshot('full_page_screenshot_38')
     global bill_recommender2
-
     # approver_id_4 = l2_page.find_approver_id()
     bill_recommender2 = str(int(l3_page.find_approver_id()))
-    print("Test 17: Vendor bill approval...{}".format(bill_recommender2))
+    print(f"Bill Recommender 2: {bill_recommender2}")    
 
     # logout from the page
     r2_page = MainNavigationBar(page)
@@ -496,20 +504,18 @@ def test_18_vendor_bill_recommender2_approval(page, new_tab):
         timeout=60000
     )
     l2_page = BillList(page)
-    bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
     l2_page.navigate_to_url(bill_payable_url)
     # l2_page.search_bill(bill_num)
     # l2_page.click_on_bill_num(bill_num)
     l2_page.search_bill(bill_num)
+
     new_page = new_tab(lambda p:l2_page.click_on_bill_num(bill_num))
     b_page = BillDetails(new_page)
     b_page.approve_bill()
     b_page.get_full_page_screenshot('full_page_screenshot_40')
-
     new_page.close()
 
     l3_page = BillList(page)
-    bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
     l3_page.navigate_to_url(bill_payable_url)
     # l2_page.search_bill(bill_num)
     l3_page.search_bill(bill_num)
@@ -517,7 +523,9 @@ def test_18_vendor_bill_recommender2_approval(page, new_tab):
     # b_page.get_full_page_screenshot('full_page_screenshot_41')
     global bill_approver_id
     bill_approver_id = str(int(l3_page.find_approver_id()))
-    print("Test 18: Vendor bill approval...{}".format(bill_approver_id))
+    print(f"Bill Approver : {bill_approver_id}")
+    # print("Test 18: Vendor bill approval...{}".format(bill_approver_id))
+    
     # logout from the page
     r2_page = MainNavigationBar(page)
     r2_page.exit()
@@ -537,21 +545,19 @@ def test_19_vendor_bill_approver_approval(page, new_tab):
     )
 
     l2_page = BillList(page)
-    bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
     l2_page.navigate_to_url(bill_payable_url)
     # l2_page.search_bill(bill_num)
     # l2_page.click_on_bill_num(bill_num)
     l2_page.search_bill(bill_num)
+    
     new_page = new_tab(lambda p:l2_page.click_on_bill_num(bill_num))
     b_page = BillDetails(new_page)
     b_page.approve_bill()
     b_page.get_full_page_screenshot('full_page_screenshot_42')
-
     new_page.close()
-    l3_page = BillList(page)
-    bill_payable_url = proj_url + "/procurementDashboard/myDashboard#!/thirdPartyBillPayable/billList"
-    l3_page.navigate_to_url(bill_payable_url)
     
+    l3_page = BillList(page)
+    l3_page.navigate_to_url(bill_payable_url)    
     l3_page.search_bill(bill_num)
     l3_page.wait_for_timeout(5000)
     bill_status=l3_page.find_bill_status()
